refactor(story-3-9): refactor FlashCard to use context, remove progress-related props, move markWordLearned logic to context, update docs
