+++ docs/business-requirements/epic-3-state-management-refactor-template/story-3-9-refactor-flashcard-template.md
@@ -6,7 +6,7 @@ Refactor the `FlashCard` component to consume Mandarin context directly, removin
 
 ## Status
 
-Planned
+Completed
 
 ## Epic Reference
